move findInsertionRange function to private API
