+++ addon/plugins/document-title-plugin/commands/insert-document-title.ts
@@ -1,5 +1,5 @@
 import { Command } from '@lblod/ember-rdfa-editor';
-import { findInsertionRange } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/article-structure-plugin/commands/insert-structure';
+import { findInsertionRange } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/_private/find-insertion-range';
 import { v4 as uuid } from 'uuid';
 
 interface InsertDocumentTitleArgs {
